requirements.txt
PySide6
nibabel
pydicom
SimpleITK
numpy
scikit-image
python-Levenshtein
